Move probe client creation code into own package for later extension, update IBC to secure version
